telegram-java-bot/telegram-bot/requirements.txt
aiogram==3.13.1
requests
sqlmodel
sqlalchemy
asyncpg
python-dateutil
PyMuPDF
dateparser
python-dotenv